+++ Fighting Game Engine/Time.cpp
@@ -15,7 +15,7 @@ void Time::Update(){
 	Time::time = glfwGetTime();
 	Time::deltaTime = Time::time - Time::lastTime;
 	Time::smoothDeltaTime = std::max(Time::smoothDeltaTime + (Time::deltaTime - Time::smoothDeltaTime)*0.01,Time::deltaTime);
-	updateRate = std::max(updateRate + (Time::deltaTime*VE_FRAME_RATE - updateRate)*0.01, Time::deltaTime*VE_FRAME_RATE);
+	updateRate = (float)std::max(updateRate + (Time::deltaTime*VE_FRAME_RATE - updateRate)*0.01, Time::deltaTime*VE_FRAME_RATE);
 	timeSinceLoad += deltaTime;
 }
